Split auth methods into their own files
